requirements.txt
imageai==3.0.3
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
pygame==2.3.0
PyOpenGL==3.1.6
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions==4.5.0
ultralytics
pyautogui
pygetwindow
requests
pyyadisk
